% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dengueSP}
\alias{dengueSP}
\title{Dummy epidemiological timeseries data of dengue in the Sao Paulo municipality, Brazil, in 2014-2023}
\format{
A data frame with 3 variables:
\describe{
  \item{epiyear}{Epidemiological year (between 2013-2024)}
  \item{epiweek}{Epidemiological week (between 0-53)}
  \item{inc}{Number of cases (integer)}
}
}
\usage{
dengueSP
}
\description{
A toy dataset for demonstrating the package functionality
}
\examples{
data(dengueSP)
head(dengueSP)
}
\keyword{datasets}
