% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endemic_channel.R
\name{endemic_channel}
\alias{endemic_channel}
\title{Computes endemic channel}
\usage{
endemic_channel(df, type = "geometric", remove_53 = T, show_plot = T)
}
\arguments{
\item{df}{A dataframe containing columns `epiyear`, `epiweek`, and `inc`.}

\item{type}{Indicates the type of mean used. Options: 'geometric' and 'arithmetic', defaults to 'geometric'}

\item{remove_53}{Indicates if 53rd epidemiological weeks should be removed, defaults to TRUE}

\item{show_plot}{Indicated whether to plot endemic channel result, defaults to TRUE}
}
\value{
A named list with:
\describe{
  \item{df_channel}{The dataframe with columns epiweek, mean, upper, lower}
  \item{plot}{A ggplot object of the endemic channel}
}
}
\description{
This function calculates the endemic channel using the geometric or arithmetic
mean, following the methodology of Bortman, 1999.
}
\examples{
\dontrun{
res <- endemic_channel(df)
res$df_channel
res$plot
}
}
