% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_endemic_channel.R
\name{plot_endemic_channel}
\alias{plot_endemic_channel}
\title{Returns plot of previously-calculated endemic channels}
\usage{
plot_endemic_channel(df)
}
\arguments{
\item{df}{A dataframe containing columns `epiweek`, `mean`, `upper` and `lower`.}
}
\value{
A ggplot2 object of the endemic channel
}
\description{
This function returns a ggplot2 object that is a plot of a previously-calculated
endemic channel.
}
\examples{
\dontrun{
p <- plot_endemic_channel(df)
print(p)
}
}
